face-recognition/requirements.txt
Flask>=2.0.0,<3.0.0
opencv-python>=4.5.0,<5.0.0
numpy>=1.21.0,<2.0.0
Werkzeug>=2.0.0,<3.0.0
pymongo>=4.0.0,<5.0.0
dnspython>=2.0.0
python-dotenv>=0.21.0
